+++ compiler/src/main/java/com/fizzed/rocker/model/BreakStatement.java
@@ -13,20 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fizzed.rocker;
-
-import com.fizzed.rocker.runtime.BreakException;
+package com.fizzed.rocker.model;
 
 /**
  *
  * @author joelauer
  */
-public interface ForIterator {
-    
-    int index();
-    
-    boolean first();
-    
-    boolean last();
-    
+public class BreakStatement extends TemplateUnit {
+
+    public BreakStatement(SourceRef sourceRef) {
+        super(sourceRef);
+    }
+
 }